[FIX] web_editor, website: fix onFocus/onBlur flow + add test
The flow has been broken multiple times across the years. Let's add a
test in all stable versions.

For the 14.0 forward-port, the test came with a fix which restores the
13.0 flow which we tried to change in 14.0... but only ended up broking
it. Indeed, with [1], the 13.0 was fixed. And, with [2], we tried to
simplify the flow... but the explanation given by the commit is just
wrong: the flow was just broken and the simplification that was to be
done in fact practically never occurs. The more the reason to add a
test, proving how the flow is working. The simplification may be done
someday but not sure it is really worth it.

For the 15.0 forward-port, another fix was needed because of [3] which
introduced a bug which makes the onBlur method be called multiple times
when hiding a whole editor hierarchy.

[1]: https://github.com/odoo/odoo/commit/71af0fc008f26386e0739e5583da63fe479438eb
[2]: https://github.com/odoo/odoo/commit/e3b545bf96d0d769c7f88723fae99b26afcb0644
[3]: https://github.com/odoo/odoo/commit/e2c5e3063f4c6091b2c65f2a0abcfb4b5ee052cf

closes odoo/odoo#79730

X-original-commit: f2a7aa9f65d5b2124fafea2e7ba7830f1e627850
Signed-off-by: Romain Derie (rde) <rde@odoo.com>
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>