[FIX] l10n_fr_fec: correct translation of journal name in FEC export
Before this commit, when exporting the FEC report, we were using the account journal
name in the initial language set. As it could be confusing for the French administration
to see names in a foreign language, we should use the user language instead to allow the
export of FEC report in French even if another default language was used.

closes odoo/odoo#79903

X-original-commit: 4f0e4b1bef4b672dec121ebd28d0e14d3385dcaf
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>