[FIX] stock_account: fix multi-company stock.quant value
On Fifo costing method, the 'value' field of a stock quant was computed using quantity_svl and value_svl of the product_id.
However the method 'with_company()' wasn't used for quantity_svl, resulting in an incorrect 'value' field of stock.quant

This commit add the use of 'with_company()' method to get quantity_svl.

OPW-2652211

closes odoo/odoo#79835

X-original-commit: 62c12a0d4f280ff43254d7bdd86c8dc07f2a5740
Signed-off-by: William Henrotin (whe) <whe@odoo.com>