[FIX] web_editor: not override root on website
Before this commit, if the editor was initiated as a simple
editor rather than the website editor (eg. the editor in the forum or profile)
the $root of the wysiwyg was wrongly assigned to #wrapwrap rather
than the $editable.

Subsequent binding (eg. LinkPopover) was bound to too many elements
which generated incorrect behaviors (eg, link popover on triggered on every
links).

Task-2691117

closes odoo/odoo#79742

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>