[FIX] project: fix following on assignation
This commit fixes an issue since user_id was changed to user_ids.

The default behaviour of `_message_auto_subscribe_followers` could not
work with the user_ids field as it is only meant to work with user_id.

This also fixes a related issue due to the same problem where the
assignation emails were not sent.

TaskId-2691486

closes odoo/odoo#79791

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>