[FIX] mail: duplicate attachement using sudo
To be able to copy an attachement, we need to be have write access on
the linked record.
In this wizard, write access on mail.template is then required which
is no longer the case since accdd440a492d68dc349a52
Read access on the template is already checked when reading the values
of the tempate.

closes odoo/odoo#79888

Signed-off-by: Olivier Dony (odo) <odo@openerp.com>