[FIX] hr_holidays: let 'new time off' button open dialog
Before this commit, if the user doesn't have super power,
he will not be able to launch the dialog to encode his time off,
because a search on ir.ui.view is done in rpc
but it is only allowed for powered user.

Now we user the method get_view_id instead.

Only applicable to 15.0

task-2694180

closes odoo/odoo#79958

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>