[FIX] stock_landed_costs: Allow to use LC without real_time
The requirement of having real_time valuation in product categories to use LC is not needed
at all, as the information is taken from stock.valuation.layer, but it was previously
enforced due to the leftovers of the previous algorithm.

We remove such requirement in this commit, and also not creating the LC journal entry in
such cases.

TT32954

closes odoo/odoo#81291

X-original-commit: 659bda6170358c294a0e145ebe86743ada52bc59
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Arnold Moyaux <arm@odoo.com>